fix: resolve package file editing and root container download issues
